Rename ModalContent component file name
